🚚 chore/Rename redux to state
