+++ client/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
-import { store } from './redux'
+import { store } from './state'
 import { ChakraProvider, ColorModeScript, extendTheme } from '@chakra-ui/react'
 
 import { App } from './components/App'
