--- client/src/redux/index.ts
@@ -1,2 +1,0 @@
-export { store, RootActionTypes } from './store'
-export type { RootState, AppDispatch, AppThunk, RootActions } from './store'
