+++ client/src/state/index.ts
@@ -1,0 +1,2 @@
+export { store, RootActionTypes, useAppSelector, useAppDispatch } from './store'
+export type { RootState, AppDispatch, AppThunk, RootActions } from './store'
